refactor flagd packaging
