+++ providers/flagd/pkg/service/rpc/mock_client_test.go
@@ -1,10 +1,9 @@
-package service
+package rpc
 
 import (
-	"context"
-
 	v1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/schema/v1"
-	"connectrpc.com/connect"
+	"context"
+	"github.com/bufbuild/connect-go"
 )
 
 // MockClient is a test mock for service client
