+++ providers/flagd/pkg/service/rpc/retryCounter.go
@@ -1,4 +1,4 @@
-package service
+package rpc
 
 import "time"
