+++ providers/flagd/pkg/service/rpc/service_eventing_test.go
@@ -1,15 +1,14 @@
-package service
+package rpc
 
 import (
+	schemaV1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/schema/v1"
 	"context"
 	"errors"
-	"testing"
-	"time"
-
-	schemaV1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/schema/v1"
-	"github.com/open-feature/go-sdk-contrib/providers/flagd/pkg/cache"
+	"github.com/open-feature/go-sdk-contrib/providers/flagd/internal/cache"
 	of "github.com/open-feature/go-sdk/pkg/openfeature"
 	"google.golang.org/protobuf/types/known/structpb"
+	"testing"
+	"time"
 )
 
 func TestRetries(t *testing.T) {
